requirements.txt
# Runtime dependencies
azure-storage-queue==12.14.1
confluent-kafka==2.6.1
croniter==6.0.0
PyYAML==6.0.3
structlog==25.4.0
httpx==0.28.1

